Replace Realm with Room; Kotlin too

+++ app/src/main/java/com/awscherb/cardkeeper/ui/card_detail/CardDetailActivity.java
@@ -2,7 +2,6 @@ package com.awscherb.cardkeeper.ui.card_detail;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v7.widget.Toolbar;
 
 import com.awscherb.cardkeeper.R;
 import com.awscherb.cardkeeper.ui.base.BaseActivity;
@@ -24,7 +23,7 @@ public class CardDetailActivity extends BaseActivity {
 
         setUpToolbar();
 
-        insertFragment(CardDetailFragment.newInstance(getIntent().getLongExtra(EXTRA_CARD_ID, 0)));
+        insertFragment(CardDetailFragment.newInstance(getIntent().getIntExtra(EXTRA_CARD_ID, 0)));
 
     }
